(#7730): Rename files away from showcase wording
