+++ www/src/views/starters/filtered-starters.js
@@ -9,7 +9,7 @@ import presets, { colors } from "../../utils/presets"
 import styles from "../shared/styles"
 
 import LHSFilter from "./lhs-filter"
-import ShowcaseList from "./showcase-list"
+import StartersList from "./starters-list"
 import Button from "../../components/button"
 import {
   SidebarHeader,
@@ -21,7 +21,7 @@ import {
 } from "../shared/sidebar"
 import ResetFilters from "../shared/reset-filters"
 
-export default class FilteredShowcase extends Component {
+export default class FilteredStarterLibrary extends Component {
   state = {
     sitesToShow: 9,
   }
@@ -184,7 +184,7 @@ export default class FilteredShowcase extends Component {
               </label>
             </div>
           </ContentHeader>
-          <ShowcaseList
+          <StartersList
             urlState={urlState}
             sortRecent={urlState.sort === `recent`}
             starters={starters}
