+++ www/src/views/starters/lhs-filter.js
